// src/app/@core/api/models/mail-request-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface MailRequestDto {
  additionalData?: {
[key: string]: any;
};
  appCode: string;
  company?: string;
  email: string;
  mailType: string;
  message: string;
  name: string;
  phone?: string;
  recaptchaToken?: string;
  subject?: string;
  website?: string;
}